refactor: claude code do refactor
initial prompt "read src directory, there quite a few problem with this project i.e. for rendering wiki the component itself become more complicated and harder to maintain. look into entier project to see if complexity can be reduced and do it (i.e. refactor components, split component into smaller files."
