+++ package.json
@@ -84,6 +84,10 @@
     "overrides": {
       "react": "19.0.0",
       "react-dom": "19.0.0"
-    }
+    },
+    "onlyBuiltDependencies": [
+      "esbuild",
+      "sharp"
+    ]
   }
 }
