+++ src/packlets/layouts/base.astro
@@ -4,7 +4,6 @@ import Blank from './blank.astro'
 import AuthButton from '$components/authButton.astro'
 import Announcement from '$components/announcement.astro'
 import FooterLogo from './footerLogo.astro'
-import HeaderLogo from './headerLogo.astro'
 
 import { getFeatureFlags } from '$functions/getFeatureFlags'
 import HeaderOgImage from './headerOgImage.astro'
