+++ src/packlets/wiki/wiki.astro
@@ -1,15 +1,11 @@
 ---
 import Head from '$commons/head.astro'
 import type { WikiPageViewModel } from '$functions/wiki/getWikiPageViewModel'
-import BaseLayout from '$layouts/base.astro'
-import ContentRenderer from '$markdown/contentRenderer.astro'
 import '@fontsource-variable/inter'
-import { Icon } from 'react-iconify-icon-wrapper'
 import WikiEditor from './editor/editor.astro'
-import WikiLinkChecker from './linkChecker'
-import WikiUpdateChecker from './updateChecker'
 import WikiView from './view.astro'
-import WikiViewShortcuts from './viewShortcuts.svelte'
+import SimpleWikiView from './components/SimpleWikiView.astro'
+import WikiContent from './components/WikiContent.astro'
 
 export interface Props {
   viewModel: WikiPageViewModel
@@ -25,7 +21,7 @@ const { head, body, perf } = Astro.props.viewModel
 />
 
 {
-  body.mode === 'editor' || body.mode === 'view' ? (
+  body.mode === 'editor' ? (
     <WikiView
       pageRef={body.pageRef}
       title={body.title}
@@ -35,58 +31,34 @@ const { head, body, perf } = Astro.props.viewModel
       lastModifiedBy={body.lastModifiedBy}
       editable={body.editable}
     >
-      {body.mode === 'editor' ? (
-        <WikiEditor pageRef={body.pageRef} file={body.file} />
-      ) : (
-        <Fragment>
-          {body.validationError ? (
-            <div class="relative mb-8 rounded-sm border border-red-400 bg-red-100 px-4 py-3 text-red-700">
-              <strong class="mr-2 font-bold">Invalid front matter:</strong>
-              <span class="block sm:inline">{body.validationError}</span>
-            </div>
-          ) : null}
-          <ContentRenderer content={body.html} pageRef={body.pageRef} />
-          <WikiUpdateChecker
-            initialContentHash={body.contentHash}
-            pageRef={body.pageRef}
-            client:idle
-          />
-          <WikiLinkChecker client:idle />
-          {body.filePath ? (
-            <WikiViewShortcuts path={body.filePath} client:only="svelte" />
-          ) : null}
-        </Fragment>
-      )}
+      <WikiEditor pageRef={body.pageRef} file={body.file} />
+    </WikiView>
+  ) : body.mode === 'view' ? (
+    <WikiView
+      pageRef={body.pageRef}
+      title={body.title}
+      headings={body.headings}
+      parsedFrontMatter={body.parsedFrontMatter}
+      lastModified={body.lastModified}
+      lastModifiedBy={body.lastModifiedBy}
+      editable={body.editable}
+    >
+      <WikiContent 
+        html={body.html} 
+        pageRef={body.pageRef} 
+        contentHash={body.contentHash}
+        filePath={body.filePath}
+        validationError={body.validationError}
+      />
     </WikiView>
   ) : (
-    <BaseLayout>
-      <div class="px-6 pb-10">
-        <div class="relative mx-auto max-w-6xl">
-          <h1 class="mb-8 text-3xl">{body.pageRef}</h1>
-          <ContentRenderer content={body.html} pageRef={body.pageRef} />
-          {body.editable && (
-            <div class="mt-8 flex">
-              <a
-                href={`/wiki/${body.pageRef}/editor`}
-                data-js="wiki-edit-link"
-                class="inline-flex items-center space-x-2"
-              >
-                <span class="edit">Edit this page</span>
-                <Icon icon="pixelarticons:edit" />
-              </a>
-            </div>
-          )}
-          <WikiUpdateChecker
-            initialContentHash={body.contentHash}
-            pageRef={body.pageRef}
-            client:idle
-          />
-          {body.filePath ? (
-            <WikiViewShortcuts path={body.filePath} client:only="svelte" />
-          ) : null}
-        </div>
-      </div>
-    </BaseLayout>
+    <SimpleWikiView
+      pageRef={body.pageRef}
+      html={body.html}
+      contentHash={body.contentHash}
+      editable={body.editable}
+      filePath={body.filePath}
+    />
   )
 }
