+++ src/pages/auth/authorize.astro
@@ -1,7 +1,6 @@
 ---
 import { IdTokenInspector } from '$components/idTokenInspector'
 import { isSensitiveScope, oauthClients } from '$constants/oauth'
-import { errorPage } from '$functions/errorPage'
 import Blank from '$layouts/blank.astro'
 
 import type { OAuthClientConfig } from '$constants/oauth'
